+++ scripts/promptlang.py
@@ -3,13 +3,16 @@ import sys
 base_dir = scripts.basedir()
 sys.path.append(base_dir)
 
-from lib_prompt_fusion import interpolation_tensor, prompt_parser as prompt_fusion_parser, hijacker, empty_cond, global_state
+from lib_prompt_fusion import hijacker, empty_cond, global_state, interpolation_tensor, geometries, interpolation_functions, prompt_parser as prompt_fusion_parser, ast_nodes
 import importlib
 importlib.reload(hijacker)
 importlib.reload(empty_cond)
 importlib.reload(global_state)
 importlib.reload(interpolation_tensor)
+importlib.reload(geometries)
+importlib.reload(interpolation_functions)
 importlib.reload(prompt_fusion_parser)
+importlib.reload(ast_nodes)
 from modules import prompt_parser, script_callbacks, shared
 import torch
 import gradio as gr
